Refactor types and interfaces
