+++ packages/tags/.size-snapshot.json
@@ -1,13 +1,13 @@
 {
   "index.cjs.js": {
-    "bundled": 11176,
-    "minified": 7554,
-    "gzipped": 2476
+    "bundled": 11194,
+    "minified": 7564,
+    "gzipped": 2485
   },
   "index.esm.js": {
-    "bundled": 9936,
-    "minified": 6596,
-    "gzipped": 2251,
+    "bundled": 9954,
+    "minified": 6606,
+    "gzipped": 2259,
     "treeshaked": {
       "rollup": {
         "code": 5582,
